+++ src/card/cards/stale_followups.rs
@@ -70,3 +70,5 @@ pub const SIMULACRUM: CardDefinitionId = CardDefinitionId(1714);
 pub const REVERSE_POLARITY: CardDefinitionId = CardDefinitionId(1715);
 pub const ALI_FROM_CAIRO: CardDefinitionId = CardDefinitionId(1716);
 pub const FORETHOUGHT_AMULET: CardDefinitionId = CardDefinitionId(1717);
+pub const LURE: CardDefinitionId = CardDefinitionId(1718);
+pub const MARBLE_PRIEST: CardDefinitionId = CardDefinitionId(1719);